Pin ruff and fix what it reports
CI's lint job had been red on every push, and not because of the code: the
dev extra asked for ruff>=0.6, so each job installed whatever release was
current, and ruff 0.16 lints far more by default than the 0.14 in use
here. Same source, two answers. Pinning makes the gate reproducible, and
what the newer ruff reported is fixed rather than silenced:

- imports sorted the way ruff now sorts them, re.IGNORECASE instead of the
  re.I alias, itertools.pairwise over zip(x, x[1:]), round() without an
  int() around it, the ytdl-format string parenthesised instead of
  implicitly concatenated, and TypeError for a value that is not a dict.
- The RTÉ scraper's blind excepts now name what can actually go wrong
  (PlaywrightError, AttributeError, RuntimeError, OSError). Probing still
  carries on to the next control, and the two silent fallbacks in the
  click path now say so on stdout, which is what they used to hide.

Verified: ruff 0.16.8 clean, 117 tests pass, and an RTÉ resolution run
still reaches the audio URL with the narrowed handlers.

+++ src/subcast/srt.py
@@ -6,7 +6,6 @@ import json
 import textwrap
 from pathlib import Path
 
-
 # Columns per caption line: one terminal line, kept short enough to read
 # without moving your eyes across the screen.
 LINE_WIDTH = 42
@@ -14,9 +13,7 @@ LINE_WIDTH = 42
 def srt_timestamp(
     seconds: float,
 ) -> str:
-    milliseconds = int(
-        round(seconds * 1000)
-    )
+    milliseconds = round(seconds * 1000)
 
     hours, milliseconds = divmod(
         milliseconds,
